# services/gateway/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EcommerceGo API Gateway",
    "description": "Unified entry point for all EcommerceGo microservices. All requests require a JWT Bearer token in the Authorization header unless noted otherwise. Prices are always in cents (int64).",
    "version": "1.0.0",
    "contact": {
      "name": "EcommerceGo",
      "url": "https://github.com/utafrali/EcommerceGo"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development (API Gateway)"
    }
  ],
  "security": [
    { "BearerAuth": [] }
  ],
...
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "tags": ["products"],
        "summary": "List products",
        "operationId": "listProducts",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } },
          { "name": "category_id", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "brand_id", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["active", "inactive", "draft"] } },
          { "name": "sort_by", "in": "query", "schema": { "type": "string", "enum": ["price_asc", "price_desc", "newest", "name_asc"] } }
        ],
        "responses": {
          "200": { "description": "Paginated product list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaginatedProductsResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "tags": ["products"],
        "summary": "Create product",
        "operationId": "createProduct",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateProductRequest" } } } },
        "responses": {
          "201": { "description": "Product created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "tags": ["products"],
        "summary": "Get product by ID",
        "operationId": "getProduct",
        "parameters": [ { "$ref": "#/components/parameters/IDParam" } ],
        "responses": {
          "200": { "description": "Product details", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "put": {
        "tags": ["products"],
        "summary": "Update product",
        "operationId": "updateProduct",
        "parameters": [ { "$ref": "#/components/parameters/IDParam" } ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateProductRequest" } } } },
        "responses": {
          "200": { "description": "Product updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "tags": ["products"],
        "summary": "Delete product",
        "operationId": "deleteProduct",
        "parameters": [ { "$ref": "#/components/parameters/IDParam" } ],
        "responses": {
          "200": { "description": "Product deleted" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/api/v1/search/": {
      "get": {
        "tags": ["search"],
        "summary": "Search products",
        "operationId": "searchProducts",
        "parameters": [
          { "name": "q", "in": "query", "schema": { "type": "string" } },
          { "name": "category_id", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "brand_id", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "min_price", "in": "query", "schema": { "type": "integer", "format": "int64", "minimum": 0 } },
          { "name": "max_price", "in": "query", "schema": { "type": "integer", "format": "int64", "minimum": 0 } },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["relevance", "price_asc", "price_desc", "newest", "name_asc", "name_desc"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } }
        ],
        "responses": {
          "200": { "description": "Search results" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/v1/auth/login or /api/v1/auth/register"
      }
    },
    "parameters": {
      "IDParam": {
        "name": "id",
        "in": "path",
        "description": "Resource UUID",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request body or parameters",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "Unauthorized": {
        "description": "Missing or invalid JWT token",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      }
    },
    "schemas": {
...
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": { "type": "string", "example": "UNAUTHORIZED" },
          "message": { "type": "string", "example": "missing or invalid token" },
          "request_id": { "type": "string", "example": "req-abc123" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "$ref": "#/components/schemas/ErrorDetail" }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "example": "Running Shoe Pro" },
          "slug": { "type": "string", "example": "running-shoe-pro" },
          "description": { "type": "string" },
          "category_id": { "type": "string", "format": "uuid" },
          "brand_id": { "type": "string", "format": "uuid" },
          "base_price": { "type": "integer", "format": "int64", "example": 9999, "description": "Price in cents" },
          "currency": { "type": "string", "example": "USD" },
          "status": { "type": "string", "enum": ["active", "inactive", "draft"] },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "ProductResponse": {
        "type": "object",
        "properties": { "data": { "$ref": "#/components/schemas/Product" } }
      },
      "PaginatedProductsResponse": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "per_page": { "type": "integer" },
          "total_pages": { "type": "integer" }
        }
      },
...
      "CreateProductRequest": {
        "type": "object",
        "required": ["name", "base_price", "currency"],
        "properties": {
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "description": { "type": "string" },
          "category_id": { "type": "string", "format": "uuid" },
          "brand_id": { "type": "string", "format": "uuid" },
          "base_price": { "type": "integer", "format": "int64", "minimum": 0, "description": "Price in cents" },
          "currency": { "type": "string", "example": "USD" },
          "status": { "type": "string", "enum": ["active", "inactive", "draft"], "default": "draft" }
        }
      },
      "UpdateProductRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "base_price": { "type": "integer", "format": "int64", "minimum": 0 },
          "status": { "type": "string", "enum": ["active", "inactive", "draft"] }
        }
      },
...
    }
  }
}